refactor: clarify manager assignment relationships
