+++ app/Models/Contracts/Manageable.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\Models\Contracts;
 
-use App\Models\Concerns\CanBeManaged;
+use App\Models\Concerns\HasManagerAssignments;
 use App\Models\Managers\Manager;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
@@ -14,7 +14,7 @@ use Illuminate\Database\Eloquent\Relations\Pivot;
  * @template TPivotModel of Pivot The pivot model for the manager relationship
  * @template TModel of Model The model that can be managed
  *
- * @see CanBeManaged For the trait implementation
+ * @see HasManagerAssignments For the trait implementation
  */
 interface Manageable
 {
