+++ tests/Unit/Models/TagTeams/TagTeamTest.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 use App\Builders\Roster\TagTeamBuilder;
 use App\Enums\Shared\EmploymentStatus;
-use App\Models\Concerns\CanBeManaged;
 use App\Models\Concerns\HasChampionshipReigns;
+use App\Models\Concerns\HasManagerAssignments;
 use App\Models\Concerns\HasMatchParticipations;
 use App\Models\Concerns\HasStableMemberships;
 use App\Models\Concerns\IsEmployable;
@@ -71,7 +71,7 @@ describe('TagTeam Model Unit Tests', function () {
 
     describe('trait integration', function () {
         test('uses all required traits', function () {
-            expect(class_uses(TagTeam::class))->toContain(CanBeManaged::class);
+            expect(class_uses(TagTeam::class))->toContain(HasManagerAssignments::class);
             expect(class_uses(TagTeam::class))->toContain(HasStableMemberships::class);
             expect(class_uses(TagTeam::class))->toContain(HasChampionshipReigns::class);
             expect(class_uses(TagTeam::class))->toContain(HasFactory::class);
